+++ routes/user.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
-const userController = require('../controller/userController')
+const userController = require('../controllers/userController.js')
+
 
 //上傳圖片
 const multer = require('multer')
